+++ src/Annotation/VirtualProperty.php
@@ -26,8 +26,9 @@ use Tebru\AnnotationReader\AbstractAnnotation;
  * You could deserialize into a class with an id property that has @VirtualProperty("data") on the
  * class. Similarly, serializing that object would produce the above json.
  *
- * The value of this annotation is only used during serialization. Please note that this will always
- * wrap classes of the specified type, regardless of where they occur in the tree.
+ * The value of this annotation acts as the serialized name. If a SerializedName annotation also exists, that will
+ * take precedence. Please note that this will always wrap classes of the specified type, regardless of where they
+ * occur in the tree.
  *
  * @author Nate Brunette <n@tebru.net>
  * @Annotation
@@ -41,5 +42,15 @@ class VirtualProperty extends AbstractAnnotation
     protected function init(): void
     {
         $this->value = $this->data['value'] ?? null;
+    }
+
+    /**
+     * Return the serialized name if it exists or null
+     *
+     * @return null|string
+     */
+    public function getSerializedName(): ?string
+    {
+        return $this->value;
     }
 }